Move fmt::Write implementation to the UEFI text output itself
