+++ uefi-logger/Cargo.toml
@@ -8,4 +8,3 @@ edition = "2018"
 [dependencies]
 uefi = { path = ".." }
 log = { version = "0.4", default-features = false }
-ucs2 = "0.1"
